MacOSX support: workspace changes and Travis changes
We no longer build glib from source, but instead require it to be
available locally.

Also modify other "local system" repositories to point at homebrew
directory locations.

The workspace changes are not strictly better or worse than the master
branch, since users would generally be expected to adjust the local
repositories anyway.

The new Travis script uses homebrew for additional library
installations. It only builds a small set of targets that work with
SDL. The //:game_lib_sdl rule is updated to link with `-framework
OpenGL`.

Also update the README.md to state the nature of this branch.